Moved the filter_genotypes function
The clunky nature of generate.py ment that the filter_genotypes function had to be placed in that file. since that is no longer the case, I moved it back to the filters.py file.
